Use a single stencil buffer for a given width,height,samplecount
Review URL: http://codereview.appspot.com/4854044/


git-svn-id: http://skia.googlecode.com/svn/trunk@2061 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gpu/include/GrTexture.h
@@ -64,7 +64,9 @@ public:
      *  Approximate number of bytes used by the texture
      */
     virtual size_t sizeInBytes() const {
-        return fAllocatedWidth * fAllocatedHeight * GrBytesPerPixel(fConfig);
+        return (size_t) fAllocatedWidth *
+                        fAllocatedHeight *
+                        GrBytesPerPixel(fConfig);
     }
 
     /**
